refactor: change action utils file name
